requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Machine Learning
scikit-learn>=1.3.0
scipy>=1.11.0

# Computer Vision
opencv-python>=4.8.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Dimensionality Reduction & Visualization
umap-learn>=0.5.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0

# Web App
streamlit>=1.28.0
streamlit-drawable-canvas>=0.9.0

# Utilities
tqdm>=4.65.0
Pillow>=10.0.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.1.0
